split the logic and styles of offer page
