+++ src/pages/OfferModal/OfferPage/index.js
@@ -1,1 +1,1 @@
-export { default } from './OfferPage';
+export { default } from './useOfferPage';
